requirements.txt
# ─────────────────────────────────────────────
# EagleEye-Nigeria — Python Dependencies
# Python 3.13+ compatible
# Install with: pip install -r requirements.txt --only-binary=:all:
# ─────────────────────────────────────────────

# ── API Framework ─────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.1

# ── Geospatial ────────────────────────────────
geopandas>=1.0.0
shapely>=2.0.6
pyproj>=3.6.0
rasterio>=1.3.10
earthengine-api>=0.1.401
folium>=0.16.0
numpy>=1.24.0

# ── Data Fetching & Processing ────────────────
requests>=2.32.0
httpx>=0.27.0
pandas>=2.0.0
numpy>=1.26.0

# ── AI / Computer Vision ──────────────────────
torch>=2.3.0
torchvision>=0.18.0
opencv-python>=4.9.0.80
Pillow>=10.3.0
scikit-learn>=1.5.0

# ── Database (event logging) ──────────────────
sqlalchemy>=2.0.30

# ── Testing ───────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.6

# ── Dev Tools ─────────────────────────────────
black>=24.4.2
ruff>=0.4.4
